+++ src/ui-component/editor/ajsFlow/FlowContents.tsx
@@ -61,26 +61,41 @@ const FlowContents: FC = () => {
     const [edges, setEdges] = useState<Edge[]>([]);
     const [dialogData, setDialogData] = useState<UnitEntity | undefined>();
 
-    const changeDodumentFn = (type: string, data: unknown) => {
-        const unitEntities: UnitEntity[] = parse(data as string)
-            .map((rootUnitOfJSON: Unit) => Unit.createFromJSON(rootUnitOfJSON))
-            .map((unit: Unit) => tyFactory(unit));
-        setUnitEntities(() => unitEntities);
-        setCurrentUnitEntity(() => undefined);
+    const theme = useMemo(() => createTheme({
+        palette: {
+            mode: isDarkMode ? 'dark' : 'light'
+        }
+    }), [isDarkMode]);
+
+    const updateNodesAndEdges = (unitEntity?: UnitEntity) => {
+        if (!unitEntity) {
+            setNodes(() => []);
+            setEdges(() => []);
+            return;
+        }
+        const { nodes, edges } = createReactFlowData(unitEntity, theme);
+        nodes.forEach(node => {
+            node.data = { ...node.data, ...dialogDataState, ...currentUnitEndityState };
+        });
+        setNodes(() => nodes);
+        setEdges(() => edges);
     };
+
+    useEffect(() => updateNodesAndEdges(currentUnitEntity), [currentUnitEntity]);
     useEffect(() => {
+        const changeDodumentFn = (type: string, data: unknown) => {
+            const unitEntities: UnitEntity[] = parse(data as string)
+                .map((rootUnitOfJSON: Unit) => Unit.createFromJSON(rootUnitOfJSON))
+                .map((unit: Unit) => tyFactory(unit));
+            setUnitEntities(unitEntities);
+            setCurrentUnitEntity(undefined);
+        };
         window.EventBridge.addCallback('changeDocument', changeDodumentFn);
         window.vscode.postMessage({ type: 'ready' });
         return () => {
             window.EventBridge.removeCallback('changeDocument', changeDodumentFn);
         }
     }, []); // fire this when mount.
-
-    const theme = useMemo(() => createTheme({
-        palette: {
-            mode: isDarkMode ? 'dark' : 'light'
-        }
-    }), [isDarkMode]);
 
     const flowMenuState = {
         menuStatus: menuStatus,
@@ -99,43 +114,26 @@ const FlowContents: FC = () => {
         setDialogData: setDialogData,
     };
 
-    const createNodesAndEdges = (unitEntity?: UnitEntity) => {
-        if (!unitEntity) {
-            setNodes(() => []);
-            setEdges(() => []);
-            return;
-        }
-        const { nodes, edges } = createReactFlowData(unitEntity, theme);
-        nodes.forEach(node => {
-            node.data = { ...node.data, ...dialogDataState, ...currentUnitEndityState };
-        });
-        setNodes(() => nodes);
-        setEdges(() => edges);
-    }
-    useEffect(
-        () => createNodesAndEdges(currentUnitEntity)
-        , [currentUnitEntity]
-    );
-
-    return <>
+    return (
         <ThemeProvider theme={theme}>
             <ReactFlowProvider>
-                <Stack direction='row' spacing={0}>
-                    {menuStatus.menuItem1 && <FlowSelector
-                        unitEntites={unitEntities}
-                        currentUnitEndityState={currentUnitEndityState}
-                        flowMenuState={flowMenuState}
-                        drawerWidthState={drawerWidthState}
-                    />}
+                <Stack direction="row" spacing={0}>
+                    {menuStatus.menuItem1 && (
+                        <FlowSelector
+                            unitEntities={unitEntities}
+                            currentUnitEntityState={currentUnitEndityState}
+                            flowMenuState={flowMenuState}
+                            drawerWidthState={drawerWidthState}
+                        />
+                    )}
                     <Stack
-                        direction='column'
-                        sx={{
-                            marginLeft: `${drawerWidth}px`,
-                        }}
+                        direction="column"
+                        sx={{ marginLeft: `${drawerWidth}px` }}
+                        flex={1}
                     >
                         <Header
-                            flowMenuState={flowMenuState}
                             currentUnitEntityState={currentUnitEndityState}
+                            flowMenuState={flowMenuState}
                             drawerWidthState={drawerWidthState}
                         />
                         <Box
@@ -151,25 +149,18 @@ const FlowContents: FC = () => {
                                 colorMode={theme.palette.mode}
                                 nodeTypes={nodeTypes}
                             >
-                                <Background
-                                    variant={BackgroundVariant.Dots}
-                                />
-                                <Controls
-                                    position="bottom-left"
-                                    showInteractive={false}
-                                />
-                                <MiniMap
-                                    pannable
-                                    zoomable
-                                />
+                                <Background variant={BackgroundVariant.Dots} />
+                                <Controls position="bottom-left" showInteractive={false} />
+                                <MiniMap pannable zoomable />
                             </ReactFlow>
-                            {dialogData
-                                && <UnitEntityDialog dialogData={dialogData} onClose={() => setDialogData(undefined)} />}
+                            {dialogData && (
+                                <UnitEntityDialog dialogData={dialogData} onClose={() => setDialogData(undefined)} />
+                            )}
                         </Box>
                     </Stack>
                 </Stack>
             </ReactFlowProvider>
         </ThemeProvider>
-    </>;
+    );
 };
 export default memo(FlowContents);